+++ packages/ckeditor5-editor-balloon/CHANGELOG.md
@@ -5,4 +5,4 @@ Changelog
 
 ### Features
 
-* The first implementation of the balloon toolbar editor. Closes [#1](https://github.com/ckeditor/ckeditor5-editor-balloon-toolbar/issues/1). ([a4462ac](https://github.com/ckeditor/ckeditor5-editor-balloon-toolbar/commit/a4462ac))
+* The first implementation of the balloon editor. Closes [#1](https://github.com/ckeditor/ckeditor5-editor-balloon/issues/1). ([a4462ac](https://github.com/ckeditor/ckeditor5-editor-balloon/commit/a4462ac))